Implementação de Segurança: Auditoria, Rate Limit e Health Checks

+++ src/app/FuncionarioRouter.py
@@ -1,6 +1,8 @@
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status, Request
 from sqlalchemy.orm import Session
 from typing import List
+from slowapi.errors import RateLimitExceeded
+from services.AuditoriaService import AuditoriaService
 
 # Domain Schemas
 from domain.schemas.FuncionarioSchema import (
@@ -15,6 +17,7 @@ from infra.orm.FuncionarioModel import FuncionarioDB
 from infra.database import get_db
 from infra.security import get_password_hash
 from infra.dependencies import get_current_active_user, require_group
+from infra.rate_limit import limiter, get_rate_limit
 
 
 router = APIRouter()
@@ -27,9 +30,11 @@ router = APIRouter()
     response_model=List[FuncionarioResponse],
     tags=["Funcionário"],
     status_code=status.HTTP_200_OK,
-    summary="Listar todos os funcionários",
-)
+    summary="Listar todos os funcionários - protegida por autenticação e grupo 1")
+@limiter.limit(get_rate_limit("moderate"))
+
 async def get_funcionarios(
+    request : Request,
     db: Session = Depends(get_db),
     current_user: FuncionarioAuth = Depends(require_group([1])),
 ):
@@ -85,6 +90,7 @@ async def get_funcionario(
 )
 async def post_funcionario(
     funcionario_data: FuncionarioCreate,
+    request : Request,
     db: Session = Depends(get_db),
     current_user: FuncionarioAuth = Depends(require_group([1])),
 ):
@@ -121,6 +127,20 @@ async def post_funcionario(
         db.commit()
         db.refresh(novo_funcionario)
 
+        db.refresh(novo_funcionario) 
+        
+        # Depois de tudo executado e antes do return, registra a ação na auditoria
+        AuditoriaService.registrar_acao(
+        db=db,
+        funcionario_id=current_user.id,
+        acao="CREATE",
+        recurso="FUNCIONARIO",
+        recurso_id=novo_funcionario.id,
+        dados_antigos=None,
+        dados_novos=novo_funcionario, # Objeto SQLAlchemy com dados novos
+        request=request # Request completo para capturar IP e user agent
+        )
+
         return novo_funcionario
 
     except HTTPException:
@@ -143,6 +163,7 @@ async def post_funcionario(
 async def put_funcionario(
     id: int,
     funcionario_data: FuncionarioUpdate,
+    request: Request,
     db: Session = Depends(get_db),
     current_user: FuncionarioAuth = Depends(require_group([1])),
 ):
@@ -173,6 +194,20 @@ async def put_funcionario(
         if funcionario_data.senha:
             funcionario_data.senha = get_password_hash(funcionario_data.senha)
 
+        # se informado grupo, valida se é um grupo válido
+        if funcionario_data.grupo:
+            if funcionario_data.grupo not in [1, 2, 3]:
+                raise HTTPException( 
+                    status_code=status.HTTP_400_BAD_REQUEST,
+                    detail="Grupo inválido. Apenas grupo 1 (Admin), 2 (Atendimento Balcão) ou 3 (Atendimento Caixa) são permitidos."
+                )
+        
+        # armazena uma copia do objeto com os dados atuais, para salvar na auditoria
+        # não pode manter referencia com funcionário, para que o auditoria possa comparar
+        # por isso a cópia do __dict__
+        dados_antigos_obj = funcionario.__dict__.copy()
+    
+
         # Atualiza apenas os campos fornecidos
         update_data = funcionario_data.model_dump(exclude_unset=True)
 
@@ -181,6 +216,18 @@ async def put_funcionario(
 
         db.commit()
         db.refresh(funcionario)
+
+        # Depois de tudo executado e antes do return, registra a ação na auditoria
+        AuditoriaService.registrar_acao(
+        db=db,
+        funcionario_id=current_user.id,
+        acao="UPDATE",
+        recurso="FUNCIONARIO",
+        recurso_id=funcionario.id,
+        dados_antigos=dados_antigos_obj, # Objeto SQLAlchemy com dados antigos
+        dados_novos=funcionario, # Objeto SQLAlchemy com dados novos
+        request=request # Request completo para capturar IP e user agent
+        )
 
         return funcionario
     except HTTPException:
@@ -197,9 +244,12 @@ async def put_funcionario(
     "/funcionario/{id}",
     status_code=status.HTTP_200_OK,
     tags=["Funcionário"],
-    summary="Remover funcionário - apenas ADMIN",
+    summary="Remover funcionário - protegida por autenticação e grupo 1",
 )
+@limiter.limit(get_rate_limit("critical"))
+
 async def delete_funcionario(
+    request: Request,
     id: int,
     db: Session = Depends(get_db),
     current_user: FuncionarioAuth = Depends(require_group([1])),
@@ -213,9 +263,28 @@ async def delete_funcionario(
                 status_code=status.HTTP_404_NOT_FOUND,
                 detail="Funcionário não encontrado",
             )
+        
+        # Impede que admin se auto-exclua
+        if current_user.id == id:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Não é possível excluir seu próprio usuário"
+            )
 
         db.delete(funcionario)
         db.commit()
+
+        # Depois de tudo executado e antes do return, registra a ação na auditoria
+        AuditoriaService.registrar_acao(
+            db=db,
+            funcionario_id=current_user.id,
+            acao="DELETE",
+            recurso="FUNCIONARIO",
+            recurso_id=funcionario.id,
+            dados_antigos=funcionario,
+            dados_novos=None,
+            request=request
+        )
 
         return {"message": "Funcionario deletado com sucesso"}
